+++ test/unit/webrtc.test.ts
@@ -201,13 +201,14 @@ describe('prepareSession', () => {
     (StanzaMediaSession as jest.Mock).mockReset();
     const client = new Client({});
     const webrtc = new WebrtcExtension(client as any, {} as any);
-    webrtc['sdpOverXmpp'] = true;
+    // webrtc['sdpOverXmpp'] = true;
 
     const sessionId = 'abc';
     const pendingSession = {
       sdpOverXmpp: true
     };
     webrtc.pendingSessions[sessionId] = pendingSession as any;
+    webrtc['sessionsMap'] = { [sessionId]: true };
 
     webrtc.proxyStatsForSession = jest.fn();
     jest.spyOn(webrtc, 'getIceTransportPolicy').mockReturnValue('all');
@@ -270,10 +271,10 @@ describe('prepareSession', () => {
     StanzaMediaSession.prototype.on = jest.fn();
     const client = new Client({});
     const webrtc = new WebrtcExtension(client as any, {} as any);
-    webrtc['sdpOverXmpp'] = true;
+    webrtc['sessionsMap']['test-id'] = true;
 
     webrtc.pendingSessions['mysid'] = {
-      sessionId: 'mysid',
+      sessionId: 'test-id',
       sdpOverXmpp: true,
       autoAnswer: false,
       id: 'myid',
@@ -664,22 +665,6 @@ describe('handlePropose', () => {
     });
 
     expect(spy).toHaveBeenCalled();
-  });
-
-  it('should track sdpOverXmpp', async () => {
-    const client = new Client({});
-    const webrtc = new WebrtcExtension(client as any, {} as any);
-    webrtc['stanzaInstance'] = getFakeStanzaClient();
-
-    webrtc['sdpOverXmpp'] = false;
-    const propose1 = { autoAnswer: false, conversationId: v4(), sessionId: v4(), sdpOverXmpp: true };
-    await webrtc['handlePropose']({ from: 'someJid', to: 'myJid', propose: propose1 });
-    expect(webrtc['sdpOverXmpp']).toBeTruthy();
-
-    webrtc['sdpOverXmpp'] = true;
-    const propose2 = { autoAnswer: false, conversationId: v4(), sessionId: v4() };
-    await webrtc['handlePropose']({ from: 'someJid', to: 'myJid', propose: propose2 });
-    expect(webrtc['sdpOverXmpp']).toBeFalsy();
   });
 });